[ui] Reuse RunConfigDialog for Runs list
