+++ js_modules/dagster-ui/packages/ui-core/src/runs/RunRoot.tsx
@@ -15,8 +15,8 @@ import {useRepositoryForRunWithParentSnapshot} from '../workspace/useRepositoryF
 
 import {AssetKeyTagCollection, AssetCheckTagCollection} from './AssetTagCollections';
 import {Run} from './Run';
-import {RunConfigDialog} from './RunConfigDialog';
 import {RUN_PAGE_FRAGMENT} from './RunFragments';
+import {RunHeaderActions} from './RunHeaderActions';
 import {RunRootTrace, useRunRootTrace} from './RunRootTrace';
 import {RunStatusTag} from './RunStatusTag';
 import {DagsterTag} from './RunTag';
@@ -155,7 +155,7 @@ export const RunRoot = () => {
               </Box>
             ) : null
           }
-          right={run ? <RunConfigDialog run={run} isJob={isJob} /> : null}
+          right={run ? <RunHeaderActions run={run} isJob={isJob} /> : null}
         />
       </Box>
       <RunById data={data} runId={runId} trace={trace} />
